rework file structure of value interface
